Bug fix: {% uni_form %} tag didn't work without a helper and helper was meant to be optional.
The `BasicNode` constructor and `get_render` methods have been adapted, so
they can really handle an optional `FormHelper` parameter. When helper is not
passed a template.Variable cannot be created, so we set self.helper to None.
This means that form_html is not rendered, so the form is rendered without
a custom layout, as a generic uni-form form.

Some of the code has been cleaned & refactored to be more readable.