Adding formsets support. Work based on akaihola formsets branch.
The branch is located at:
https://github.com/akaihola/django-uni-form/commits/formsets

I'm not merging this branch for 2 reasons:
- I don't want to complicate `uni_form_tags.py` with a more object oriented
architecture, that makes it harder to follow. I don't see a good reason for
adding more inheritance for solving this problem.
- I don't want to add more tags unless necessary. So rather than having a
`{% uni_form_set %}` tag and `|as_uni_formset` filter, I have redone the
`{% uni_form %}` tag to be smarter, so it can be used for forms & formsets.
This way testing is kept simpler also.