Dropping support for older versions than Django 1.2. Removing `{% uni_form_jquery %}` tag and its template `uni_form_jquery.html`.
This is based on the discussion by pydanny and maraujop in issue #61:
https://github.com/pydanny/django-uni-form/issues#issue/61

All the lines of code that were never executed have been deleted, the
`Toggle` class has been removed and the `BaseInput` has been moved to
the helpers file, therefore we don't need a util.py file anymore.

The lines of code within the `BasicNode.get_response_dict` that supported
the toggling of fields have been removed, as the jquery tag is no longer
there.