Adding helper attribute `form_style`. This can be set to 'default' or 'inline'. Default value is 'default'.
This has been added so that setting a uni-form style for a form with a layout is
easier to do. Until now the default style was 'default' and in order to set 'inline' style
you had to set `css_class` to 'inlineLabels' manually in every Fieldset.

Tests have been updated, though some extra tests for this feature would be good to have.