Fixing many flaws in formsets support before 0.8.0 beta release.
- Formsets were all styled inline. Now you can use form_style in your formset's forms
to set your preferred style.
- Formsets got an obligatory legend that said "* Required Fields", it's been removed.
The error titles now are template variables `{{ formset_error_title }}` and `{{ form_error_title }}`
for `formset.non_form_errors` and `form.non_field_errors` respectively.
TODO: Have an easy way to set up these variables.
- Until now for using a helper for the formset's forms, it needed to be called helper
and be a variable of the form. This was hardcoded, now you can pass a helper for
rendering those forms and the formset using:
{% uni_form formset helper %}
- The `formset.management_form` was forgotten from the template.
- There was no way to avoid the formset having a <form> tag. Now the helper attributes
are used for rendering the formset general part, and the layout for rendering the
formset's forms.
- The CSRF token was repeated in every formset's form.
- The `formset.non_form_errors` were missing in some uase cases.

Formset's tests have been improved.