Removing i18n tags. Removing uni_form_tags where possible, using includes instead for making context available in the included template.
This way if `form_error_title` or `formset_error_title` are context variables,
they will be available to the included templates, while using filters this wasn't possible.
At the same time we keep DRY, like using filters.

Templates might be a little bit harder to follow now if somebody is not familiar with its
hierarchy.