Moving @issackelly and @johnthedebs work into a separate template called betterform
After consideration, having as_uni_form filter being able to render this type of
forms, doesn't look lie a very common case. ModelAdmin forms are not easy to override
for using this technique. As BetterForms by @carljm already has a render filter that can be
passed a template as an argument, I'm moving the template proposed to a separate file.

This way you can use this template to render BetterForms using django-uni-form.