Turning Input objects into Layout objects, adding a ButtonHolder Layout object.
Now we can have all type of inputs within the Layout or use `add_input` if we prefer it.

This gives us some nice extra power, we can have other Layout objects next to
the submit buttons. Also submit button don't have to be the last thing in the form.
See this example Layout:

    layout = Layout(
        Fieldset(
            u'User information',
            'username',
            'pasword',
		),
        ButtonHolder(
            HTML("""
                {% if message %}
                    <span class="notification">{{ message }}</span>
                {% endif %}"""
            ),
            Submit('login', 'Login', css_class='button white'),
            HTML("<a href='{% url forget_password1 %}'>did you forget your password?</a>"),
        ),
    )
    helper.add_layout(layout)