Refactoring `Column` and `Row` objects to inherit from a base Layout object called `Div`
If a `css_class` is specified for `Column` or `Row` it is added to its default class.
Example:

    Row('field1', 'field2', css_class='rows')

This produces a div with its class set to `class="formRow rows"`

Now we can Use Div to wrap fields and set the div id and class using `css_id` and
`css_class` respectively.