Use form's Meta information if available or Layout otherwise
This was discussed in issue #GH-88, way off track, but necessary.

`rendered_fields` is now a set, which speeds up things as we don't care
about the order of the rendered_fields but only if they were rendered
or not.

If form has Meta information `fields` or `exclude` we use them, to
render those fields, if they were not explicitly mentioned in the layout.
Otherwise we render the user's layout. This way we don't force the user
to add `fields` or `exclude` form form's Meta classes.