Adding support for hiding fields using Field layout object
Field('field_name', type='hidden')

This solves issue #GH-55

When Django calls `BoundField.as_widget`, that one calls `widget.render`
method, which builds the list of attributes calling `build_attrs`. This
one here, gives priority to the field attributes, before to the
attributes passed to the widget.

So we need to play this one different.