Adding an `attrs` attribute to FormHelper
We can use this for setting any form attribute we need. This solves #GH-48.

    form.helper.attrs = {'id': 'special-form', 'autocomplete': 'off'}

FormHelper has been refactored to encapsulate `form_id`, `form_class`
and `form_action` in an `attrs` dictionary, which is flattened and
passed to `BasicNode.get_response_dict` as `flat_attrs`. This is
rendered in the templates using |safe.

* `get_response_dict` has been refactored a little bit
* When no helper is provided, an empty one is created for handling
defaults in a easier way. Before they were hardcoded in the templates.
* `form_id`, `form_class` and `form_action` are no longer provided in
the template's context.
* Updating tests to work with uni_form TEMPLATE PACK. Adding a test for
`FormHelper.attrs`.