Simplifying Meta fields usage
When we have an instance of a form that has a Meta class, by the time we
manage it, Django has done the job of selecting which fields the user
wants, so `form.fields` is actually what the user wants to render.

Refactoring test for this case a little bit.