Added support for rendering multiple forms in the same layout
Added a `MultipleFormRenderer` class, that contains a mapping between
form names and forms, that it uses to resolve layout field names when
rendering them.

It passes the renderer object itself to `render_field_multiple` so that
it can resolve the form name when reaching a field name.

Implementation is a draft, API is subject to change.