Bug fix for model formsets and inline formsets
When rendering a formset using a FormHelper with a layout set, hidden
fields were not being rendered unless explicitly mentioned in the
layout. Using the new `render_hidden_fields` attribute in helper, we can
now make sure all hidden fields of all formset's forms are rendered.

This could fix #143, or be a different bug.

Added a test case for avoiding regressions.