Adding an aditional formset template, thanks to @runekaagaard
This template is heavily inspired by @runekaagard work in
django-crispy-forms-fancy-formsets. A great application for rendering
formsets and handling them in the client using javascript.

By default this template will render a formset within a table with
inline fields. This can be used thanks to the new and shiny helper
custom helper template attribute.

    author = Author.objects.get(pk=id_author)
    BookInlineFormSet = inlineformset_factory(Author, Book, extra=2)
    formset = BookInlineFormSet(request.POST, instance=author)
    helper = FormHelper()
    helper.template = 'bootstrap/table_inline_formset.html'

Then within template:

    {% crispy formset helper %}