Add .row class to allow custom widths on bootstrap4 form-groups
In bootstrap3 .form-group included the .row class implicitly, but in
bootstrap4 it no longer includes .row, so I've added the .row class to
all form-group div's that allow labels and fields within the component
that could potentially have custom widths.

I don't know if this is the ideal solution, the other option is warning
users that when migrating from 3 to 4 they'll need to add .row to the
wrapper_class or field.css_classes.