Drop django 1.7 support
Django 1.7 support was dropped with the release of Django 1.9, and
now we're testing all Python versions that should be supported by
Django 1.9 and 1.10.

I guess maybe we should do 2.7, 3.4, and 3.5 for Django 1.9, but
only test the minimum versions necessary for django-latest.

Source https://docs.djangoproject.com/en/dev/faq/install/#what-python-version-can-i-use-with-django