Support several bootstrap sizes in label_class for checkbox input
For checkbox input label is displayed after the input
so the input have to be aligned with the rest of the form
using col-<device>-offset-<size> class. But sometimes form may have
different sizes on different devices so we have to build offsets
for each pair of device and size.