Django 4.0 Support (#1184)
* Added USE_TZ to test settings
USE_TZ will default to True in Django 5.0 and has been the included in the default settings file since Django 1.4
https://docs.djangoproject.com/en/dev/releases/4.0/#time-zone-support

* Removed 'for ="..."' from RadioSelect's <label>.

Mirrors change in Django 4.0
https://github.com/django/django/commit/b9e872b59329393f615c440c54f632a49ab05b78

* Made is_radioselect more strict

In Django 4.0 `CheckboxSelectMultiple` is now a subclass of `RadioSelect`.
Therefore need to make is_radioselect more strickt so checkboxes do not pass as radio inputs. See
https://github.com/django/django/commit/b9e872b59329393f615c440c54f632a49ab05b78

* Started to update docs etc

* Added note to readme about Django 4.0 support

* Bumped version and updated tox.ini

* Updated publish workflow to use tox for tests

* Updated release date