Changed `get_rendered_fields()` to return SafeString. (#1271)
This function calls `render_field()` for each field which always returns a `SafeString` however, due to the join with `""` this converts it to a `str`. Let's preserve the `SafeString` which will help `|safe` calls in templates to be removed.

`render_field()` will render a field via `Template.render()` (which returns a `SafeString`) or will call a `LayoutObjects` render method which should also return a `SafeString` via either calling `render_field()` again or calling `render_to_string()` which also returns a `SafeString`.