Improved `get_layout_objects()` to return a list of Pointers.
The current implementation returns a list of lists.

Each list has two items:
- A list of locatons (int)
- The name of the object/field. (str)

This PR replaces the list with a dataclass to improve the readability of the code and to help when adding type hints.

That is the current type is Union[List[int], str]. Accessng the first item can therefore by either of these types. However, we know the type is specifically List[int]. Using a dataclass allows each compent to by typed individually.